Change ulp variable to be of type size_t

+++ rclcpp/src/rclcpp/node_interfaces/node_parameters.cpp
@@ -191,7 +191,7 @@ __lockless_has_parameter(
 // see https://en.cppreference.com/w/cpp/types/numeric_limits/epsilon
 RCLCPP_LOCAL
 bool
-__are_doubles_equal(double x, double y, int ulp = 100)
+__are_doubles_equal(double x, double y, size_t ulp = 100)
 {
   return std::abs(x - y) <= std::numeric_limits<double>::epsilon() * std::abs(x + y) * ulp;
 }
